Figures collect for both human-bot comparison and human-GPT2 comparison
I also wrote an additional script for reporting cohen's d statistic for each of the features. The comparison between results of GPT2 and bots as a whole may be valuable for our discussion, just a thought. As expected, the tests show stronger results for statistically differentiating humans against all bots, but also shows things like how much better GPT2 is at not overusing functions words like we talked about on call and how much more even the contribution of distance is between each feature for the extended dataset over all bots. Regardless of the case, these results contextualized alongside classification performance measures I think make for a good discussion even if the differentiation is not anything crazy; it's very evidently there, just not that strong